create new util class
create new util class

+++ src/main/java/edu/challenge/beat/model/Position.java
@@ -1,6 +1,5 @@
 package edu.challenge.beat.model;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
@@ -8,7 +7,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class Position {
 
-    private final int rideId;
+    private final long rideId;
     private final double latitude;
     private final double longitude;
     private final long timestamp;
